player enitity, collision, movement and renames
1. added Player Entity which extends GEntity
2. added support for entity-to-entity and entity-to-map elements collision
3. added support for accelerated movement functions
4. added support for camera focus on character
5. added support for three different Rendering functions - consecutive tilesheet images, selective tilesheet images, and static image.

+++ GCamera.cpp
@@ -49,7 +49,7 @@ void GCamera::SetPos(int X, int Y)
 }
 
 
-void GCamera::SetTarget(int* X, int* Y)
+void GCamera::SetTarget(float* X, float* Y)
 {
     targetX = X;
     targetY = Y;
